refactor knot
